Add a abstract class as the base class for all optimization related portfolio constructions.
